requirements.txt
bcrypt==4.2.1
psycopg2-binary==2.9.10
